refactor: component로 이동
