+++ components/DatepickerField.tsx
@@ -8,7 +8,7 @@ import React from 'react'
 import DatePicker, { registerLocale } from 'react-datepicker'
 import { type Control, Controller } from 'react-hook-form'
 
-import { classNames } from '../../../utils/classNames'
+import { classNames } from '../utils/classNames'
 
 interface DatePickerFieldProps {
   control: Control<any>
