+++ components/StateLabel.tsx
@@ -1,4 +1,4 @@
-import { classNames } from '../../../utils/classNames'
+import { classNames } from '../utils/classNames'
 
 export enum State {
   INUSE = 'INUSE',
