+++ pages/add.tsx
@@ -2,11 +2,11 @@ import { Noto_Sans } from 'next/font/google'
 import { type PropsWithChildren } from 'react'
 import { useForm } from 'react-hook-form'
 
-import Button from '@/features/common/components/Button'
-import ComboBox from '@/features/common/components/ComboBox'
-import DatePickerField from '@/features/common/components/DatepickerField'
-import Input from '@/features/common/components/Input'
-import Navbar from '@/features/common/components/Navbar'
+import Button from '@/components/Button'
+import ComboBox from '@/components/ComboBox'
+import DatePickerField from '@/components/DatepickerField'
+import Input from '@/components/Input'
+import Navbar from '@/components/Navbar'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
