+++ pages/calendar.tsx
@@ -1,9 +1,9 @@
 import { Noto_Sans } from 'next/font/google'
 import { useState } from 'react'
 
-import Navbar from '@/features/common/components/Navbar'
-import Calendar from '@/features/products/components/Calendar'
-import CalendarItem from '@/features/products/components/CalendarItem'
+import Navbar from '@/components/Navbar'
+import Calendar from '@/components/products/Calendar'
+import CalendarItem from '@/components/products/CalendarItem'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
