+++ pages/detail.tsx
@@ -2,9 +2,9 @@ import { Noto_Sans } from 'next/font/google'
 import Link from 'next/link'
 import { Line } from 'rc-progress'
 
-import Button from '@/features/common/components/Button'
-import Navbar from '@/features/common/components/Navbar'
-import { State, StateLabel } from '@/features/common/components/StateLabel'
+import Button from '@/components/Button'
+import Navbar from '@/components/Navbar'
+import { State, StateLabel } from '@/components/StateLabel'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
