+++ pages/edit-state.tsx
@@ -2,7 +2,7 @@ import { Noto_Sans } from 'next/font/google'
 import React from 'react'
 import { useForm } from 'react-hook-form'
 
-import Navbar from '@/features/common/components/Navbar'
+import Navbar from '@/components/Navbar'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
