+++ pages/login.tsx
@@ -6,9 +6,9 @@ import { useForm } from 'react-hook-form'
 
 import { type PostLoginProps } from '@/apis/auth/postLogin'
 import { type ServerErrorRes } from '@/apis/client'
+import Button from '@/components/Button'
+import Input from '@/components/Input'
 import { useLogin } from '@/features/auth/mutations/useLogin'
-import Button from '@/features/common/components/Button'
-import Input from '@/features/common/components/Input'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
