+++ pages/register.tsx
@@ -2,15 +2,15 @@ import { Noto_Sans } from 'next/font/google'
 import { forwardRef, type InputHTMLAttributes, useState } from 'react'
 import { useForm } from 'react-hook-form'
 
+import Button from '@/components/Button'
+import DatePickerField from '@/components/DatepickerField'
+import Input from '@/components/Input'
+import InputLabel from '@/components/InputLabel'
+import Navbar from '@/components/Navbar'
 import { useCheckLoginId } from '@/features/auth/mutations/useCheckLoginId'
 import { useEmailCode } from '@/features/auth/mutations/useEmailCode'
 import { useSignUp } from '@/features/auth/mutations/useSignUp'
 import { useVerifyEmail } from '@/features/auth/mutations/useVerifyEmail'
-import Button from '@/features/common/components/Button'
-import DatePickerField from '@/features/common/components/DatepickerField'
-import Input from '@/features/common/components/Input'
-import InputLabel from '@/features/common/components/InputLabel'
-import Navbar from '@/features/common/components/Navbar'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
